backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.1
langchain==0.0.283
google-generativeai==0.4.5
requests==2.31.0
python-dotenv==1.0.0
pydantic==1.10.11
httpx==0.24.1
google-search-results==2.4.2
